refactor: pets and todos
