+++ apps/fe-next/src/app/todos/page.tsx
@@ -1,15 +1,15 @@
 import { Suspense } from 'react';
 
-import TodosSsr from '../../components/Todos/TodosSsr';
+import TodosListSsr from '../../components/TodosList/TodosListSsr';
 import LoadingList from '../../components/LoadingList';
-import Todos from '../../components/Todos';
+import Todos from '../../components/TodosList';
 
 const TodosPage = async () => {
   return (
     <>
       <h1 className="text-3xl font-bold underline text-center">Todos</h1>
       <Suspense fallback={<LoadingList />}>
-        <TodosSsr />
+        <TodosListSsr />
       </Suspense>
       <Todos />
     </>
